Update dependencies, drop record-extra and typelevel-prelude
This means that `bower install` in this repo now succeeds again. It also
allows this project to build with purescript 0.13.0. I've dropped the
typelevel-prelude dependency because we were only using re-exports from
that. I've also dropped the record-extra dependency and used `rgb`
directly for constructing colors so that we don't need the `mapRecord`
call, as this is slightly safer (and doesn't require us to wait for
record-extra to relax its upper bounds).